bugfix: broken succinct vector for larger data
